+++ apps/web/app/api/google/threads/validation.ts
@@ -1,0 +1,9 @@
+import { z } from "zod";
+
+export const threadsQuery = z.object({
+  fromEmail: z.string().nullish(),
+  limit: z.coerce.number().max(100).nullish(),
+  type: z.string().nullish(),
+  q: z.string().nullish(),
+});
+export type ThreadsQuery = z.infer<typeof threadsQuery>;
